% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.r
\name{create_hc_plt}
\alias{create_hc_plt}
\title{Generating a plot of hc, industrialization, and a categorial variable.}
\usage{
create_hc_plt(x_var, y_var, color_var, plt_title, x_var_title)
}
\arguments{
\item{x_var}{A numeircal data.table column.}

\item{y_var}{A numerical data.table column.}

\item{color_var}{A categorial data.table column.}

\item{plt_title}{The plot's title.}

\item{x_var_title}{X-axis title.}
}
\value{
A plot presenting the correlation between x_var and y_var given the different values of color_var.
}
\description{
Generating a plot of hc, industrialization, and a categorial variable.
}
\examples{
myfunc("urban_share", "birth_rate", "Region", "Urbanization and birth rates", "Urban share")
}
